requirements.txt
flask
pillow
torch
torchvision
torchaudio
transformers
requests